Exclude ExampleFunctionTemplate_fetch when leak checking on macOS
